+++ .gitignore
@@ -1,0 +1,4 @@
+env/
+.vscode/
+.pytest_cache/
+__pycache__/
