--- requirement.txt
@@ -1,15 +1,0 @@
-atomicwrites==1.1.5
-attrs==18.1.0
-click==6.7
-Flask==1.0.2
-itsdangerous==0.24
-Jinja2==2.10
-MarkupSafe==1.0
-more-itertools==4.3.0
-pathlib2==2.3.2
-pkg-resources==0.0.0
-pluggy==0.7.1
-py==1.5.4
-pytest==3.7.1
-six==1.11.0
-Werkzeug==0.14.1
