+++ run.py
@@ -1,8 +1,8 @@
 import os
 
-from app import create_app
-config_name = os.getenv('APP_SETTINGS') # config_name = "development"
-app = create_app(config_name)
+from app import app
+# config_name = os.getenv('APP_SETTINGS') # config_name = "development"
+# app = create_app(config_name)
 
 if __name__ == '__main__':
     app.run(port=5000)
